Playlists browser: fix signals, static dependencies, menu
 * Make the Playlist Library the real source of truth
 * Add `playlists` to librarian too
 * Update audiofile to use the playlist library `playlists_featuring()` (no longer in the collection, which was ugly)
 * Sort out signals and handlers in the playlists browser to work fully (hopefully...)
 * Fix up the songsmenu, neater now
 * Split some tests out
