+++ quodlibet/ext/events/trayicon/menu.py
@@ -8,7 +8,6 @@
 
 from gi.repository import GObject, Gtk
 
-from quodlibet.browsers.playlists import PlaylistsBrowser
 from quodlibet.browsers.playlists.menu import PlaylistMenu
 
 from quodlibet import _
@@ -193,13 +192,10 @@ class IndicatorMenu(Gtk.Menu):
 
     def _new_playlist_submenu_for(self, song):
         submenu = self._playlists_item.get_submenu()
+        pl_lib = self._app.library and self._app.library.playlists
         if submenu:
             submenu.destroy()
-        playlist_menu = PlaylistMenu([song], PlaylistsBrowser.playlists())
-
-        def on_new(widget, playlist):
-            PlaylistsBrowser.changed(playlist)
-        playlist_menu.connect('new', on_new)
+        playlist_menu = PlaylistMenu([song], pl_lib)
         self._playlists_item.set_submenu(playlist_menu)
         self._playlists_item.set_sensitive(bool(song) and song.can_add)
         self._playlists_item.show_all()
