+++ tests/test_library_libraries.py
@@ -4,22 +4,14 @@
 # (at your option) any later version.
 import os
 import shutil
-from pathlib import Path
 
 from gi.repository import Gtk
 
-from quodlibet import config
 from quodlibet.formats import AudioFile
-from quodlibet.formats import AudioFileError
 from quodlibet.library.base import (Library, iter_paths, PicklingMixin)
-from quodlibet.library.file import FileLibrary
-from quodlibet.library import SongLibrary, SongFileLibrary
-from quodlibet.library.album import AlbumLibrary
 from quodlibet.util import connect_obj, is_windows
-from quodlibet.util.path import normalize_path
 from senf import fsnative
-from tests import TestCase, get_data_path, mkstemp, mkdtemp, skipIf
-from .helper import capture_output, get_temp_copy
+from tests import TestCase, mkstemp, mkdtemp, skipIf
 
 
 class Fake(int):
@@ -311,409 +303,6 @@ class TPicklingMixin(TestCase):
                 assert v.key == v2.key
         finally:
             os.unlink(filename)
-
-
-class TSongLibrary(TLibrary):
-    Fake = FakeSong
-    Frange = staticmethod(FSrange)
-    Library = SongLibrary
-
-    def test_rename_dirty(self):
-        self.library.dirty = False
-        song = self.Fake(10)
-        self.library.add([song])
-        self.failUnless(self.library.dirty)
-        self.library.dirty = False
-        self.library.rename(song, 20)
-        self.failUnless(self.library.dirty)
-
-    def test_rename(self):
-        song = self.Fake(10)
-        self.library.add([song])
-        self.library.rename(song, 20)
-        while Gtk.events_pending():
-            Gtk.main_iteration()
-        self.failUnless(song in self.changed)
-        self.failUnless(song in self.library)
-        self.failUnless(song.key in self.library)
-        self.failUnlessEqual(song.key, 20)
-
-    def test_rename_changed(self):
-        song = self.Fake(10)
-        self.library.add([song])
-        changed = set()
-        self.library.rename(song, 20, changed=changed)
-        self.assertEqual(len(changed), 1)
-        self.assertTrue(song in changed)
-
-    def test_tag_values(self):
-        self.library.add(self.Frange(30))
-        del self.added[:]
-        self.failUnlessEqual(
-            sorted(self.library.tag_values(10)), list(range(10)))
-        self.failUnlessEqual(sorted(self.library.tag_values(0)), [])
-        self.failIf(self.changed or self.added or self.removed)
-
-
-class TFileLibrary(TLibrary):
-    Fake = FakeSongFile
-    Library = FileLibrary
-
-    def test_mask_invalid_mount_point(self):
-        new = self.Fake(1)
-        self.library.add([new])
-        self.failIf(self.library.masked_mount_points)
-        self.failUnless(len(self.library))
-        self.library.mask("/adsadsafaf")
-        self.failIf(self.library.masked_mount_points)
-        self.library.unmask("/adsadsafaf")
-        self.failIf(self.library.masked_mount_points)
-        self.failUnless(len(self.library))
-
-    def test_mask_basic(self):
-        new = self.Fake(1)
-        self.library.add([new])
-        self.failIf(self.library.masked_mount_points)
-        self.library.mask(new.mountpoint)
-        self.failUnlessEqual(self.library.masked_mount_points,
-                             [new.mountpoint])
-        self.failIf(len(self.library))
-        self.failUnlessEqual(self.library.get_masked(new.mountpoint), [new])
-        self.failUnless(self.library.masked(new))
-        self.library.unmask(new.mountpoint)
-        self.failUnless(len(self.library))
-        self.failUnlessEqual(self.library.get_masked(new.mountpoint), [])
-
-    def test_remove_masked(self):
-        new = self.Fake(1)
-        self.library.add([new])
-        self.library.mask(new.mountpoint)
-        self.failUnless(self.library.masked_mount_points)
-        self.library.remove_masked(new.mountpoint)
-        self.failIf(self.library.masked_mount_points)
-
-    def test_content_masked(self):
-        new = self.Fake(100)
-        new._mounted = False
-        self.failIf(self.library.get_content())
-        self.library._load_init([new])
-        self.failUnless(self.library.masked(new))
-        self.failUnless(self.library.get_content())
-
-    def test_init_masked(self):
-        new = self.Fake(100)
-        new._mounted = False
-        self.library._load_init([new])
-        self.failIf(self.library.items())
-        self.failUnless(self.library.masked(new))
-
-    def test_load_init_nonmasked(self):
-        new = self.Fake(200)
-        new._mounted = True
-        self.library._load_init([new])
-        self.failUnlessEqual(list(self.library.values()), [new])
-
-    def test_reload(self):
-        new = self.Fake(200)
-        self.library.add([new])
-        changed = set()
-        removed = set()
-        self.library.reload(new, changed=changed, removed=removed)
-        self.assertTrue(new in changed)
-        self.assertFalse(removed)
-
-    def test_move_root(self):
-        # TODO: mountpoint tests too
-        self.library.filename = "moving"
-        root = Path(normalize_path(mkdtemp(), True))
-        other_root = Path(normalize_path(mkdtemp(), True))
-        new_root = Path(normalize_path(mkdtemp(), True))
-        in_song = FakeAudioFile(str(root / "in file.mp3"))
-        in_song.sanitize()
-        out_song = FakeAudioFile(str(other_root / "out file.mp3"))
-        # Make sure they exists
-        in_song.sanitize()
-        out_song.sanitize()
-        assert Path(in_song("~dirname")) == root, "test setup wrong"
-        assert Path(out_song("~dirname")) == other_root, "test setup wrong"
-        self.library.add([out_song, in_song])
-
-        # Run it by draining the generator
-        list(self.library.move_root(root, str(new_root)))
-        msg = f"Dir wasn't updated in {root!r} -> {new_root!r} for {in_song.key}"
-        assert Path(in_song("~dirname")) == new_root, msg
-        assert Path(in_song("~filename")) == (new_root / "in file.mp3")
-        assert Path(out_song("~dirname")) == other_root, f"{out_song} was wrongly moved"
-        assert in_song._written, "Song wasn't written to disk"
-        assert not out_song._written, "Excluded songs was written!"
-
-
-class TSongFileLibrary(TSongLibrary):
-    Fake = FakeSongFile
-    Frange = staticmethod(FSFrange)
-    Library = SongFileLibrary
-
-    def test__load_exists_invalid(self):
-        new = self.Fake(100)
-        new._valid = False
-        changed, removed = self.library._load_item(new)
-        self.failIf(removed)
-        self.failUnless(changed)
-        self.failUnless(new._valid)
-        self.failUnless(new in self.library)
-
-    def test__load_not_exists(self):
-        new = self.Fake(100)
-        new._valid = False
-        new._exists = False
-        changed, removed = self.library._load_item(new)
-        self.failIf(removed)
-        self.failIf(changed)
-        self.failIf(new._valid)
-        self.failIf(new in self.library)
-
-    def test__load_error_during_reload(self):
-        try:
-            from quodlibet import util
-            print_exc = util.print_exc
-            util.print_exc = lambda *args, **kwargs: None
-            new = self.Fake(100)
-
-            def error():
-                raise AudioFileError
-
-            new.reload = error
-            new._valid = False
-            changed, removed = self.library._load_item(new)
-            self.failUnless(removed)
-            self.failIf(changed)
-            self.failIf(new._valid)
-            self.failIf(new in self.library)
-        finally:
-            util.print_exc = print_exc
-
-    def test__load_not_mounted(self):
-        new = self.Fake(100)
-        new._valid = False
-        new._exists = False
-        new._mounted = False
-        changed, removed = self.library._load_item(new)
-        self.failIf(removed)
-        self.failIf(changed)
-        self.failIf(new._valid)
-        self.failIf(new in self.library)
-        self.failUnless(self.library.masked(new))
-
-    def __get_file(self):
-        return get_temp_copy(get_data_path('empty.flac'))
-
-    def test_add_filename(self):
-        config.init()
-        try:
-            filename = self.__get_file()
-            ret = self.library.add_filename(filename)
-            self.failUnless(ret)
-            self.failUnlessEqual(len(self.library), 1)
-            self.failUnlessEqual(len(self.added), 1)
-            ret = self.library.add_filename(filename)
-            self.failUnless(ret)
-            self.failUnlessEqual(len(self.added), 1)
-            os.unlink(filename)
-
-            filename = self.__get_file()
-            ret = self.library.add_filename(filename, add=False)
-            self.failUnless(ret)
-            self.failIf(ret in self.library)
-            self.failUnlessEqual(len(self.added), 1)
-            self.library.add([ret])
-            self.failUnless(ret in self.library)
-            self.failUnlessEqual(len(self.added), 2)
-            self.failUnlessEqual(2, len(self.library))
-            os.unlink(filename)
-
-            with capture_output():
-                ret = self.library.add_filename("")
-            self.failIf(ret)
-            self.failUnlessEqual(len(self.added), 2)
-            self.failUnlessEqual(len(self.library), 2)
-
-        finally:
-            config.quit()
-
-    def test_contains_filename(self):
-        filename = self.__get_file()
-        try:
-            assert not self.library.contains_filename(filename)
-            assert self.library.add_filename(filename, add=False)
-            assert not self.library.contains_filename(filename)
-            assert self.library.add_filename(filename)
-            assert self.library.contains_filename(filename)
-        finally:
-            os.unlink(filename)
-
-    def test_add_filename_normalize_path(self):
-        if not os.name == "nt":
-            return
-
-        config.init()
-        filename = self.__get_file()
-
-        # create a equivalent path different from the original one
-        if filename.upper() == filename:
-            other = filename.lower()
-        else:
-            other = filename.upper()
-
-        song = self.library.add_filename(filename)
-        other_song = self.library.add_filename(other)
-        self.assertTrue(song is other_song)
-        os.unlink(filename)
-        config.quit()
-
-
-class TAlbumLibrary(TestCase):
-    Fake = FakeSong
-    Frange = staticmethod(ASrange)
-    UnderlyingLibrary = Library
-
-    def setUp(self):
-        self.underlying = self.UnderlyingLibrary()
-        self.added = []
-        self.changed = []
-        self.removed = []
-
-        self._sigs = [
-            connect_obj(self.underlying, 'added', list.extend, self.added),
-            connect_obj(self.underlying,
-                        'changed', list.extend, self.changed),
-            connect_obj(self.underlying,
-                        'removed', list.extend, self.removed),
-        ]
-
-        self.library = AlbumLibrary(self.underlying)
-
-        # Populate for every test
-        self.underlying.add(self.Frange(12))
-
-    def tearDown(self):
-        for s in self._sigs:
-            self.underlying.disconnect(s)
-        self.underlying.destroy()
-        self.library.destroy()
-
-    def test_get(self):
-        key = self.underlying.get("file_1.mp3").album_key
-        self.failUnlessEqual(self.library.get(key).title, "Album 1")
-        album = self.library.get(key)
-        self.failUnlessEqual(album.key, key)
-        self.failUnlessEqual(len(album.songs), 4)
-
-        key = self.underlying.get("file_2.mp3").album_key
-        self.failUnlessEqual(self.library.get(key).title, "Album 2")
-
-    def test_getitem(self):
-        key = self.underlying.get("file_4.mp3").album_key
-        self.failUnlessEqual(self.library[key].key, key)
-
-    def test_keys(self):
-        self.failUnless(len(self.library.keys()), 3)
-
-    def test_has_key(self):
-        key = self.underlying.get("file_1.mp3").album_key
-        self.failUnless(self.library.has_key(key))
-
-    def test_misc_collection(self):
-        self.failUnless(self.library.values())
-
-    def test_items(self):
-        self.failUnlessEqual(len(self.library.items()), 3)
-
-    def test_items_2(self):
-        albums = self.library.values()
-        self.failUnlessEqual(len(albums), 3)
-        songs = self.underlying._contents.values()
-        # Make sure "all the songs' albums" == "all the albums", roughly
-        self.failUnlessEqual({a.key for a in albums},
-                             {s.album_key for s in songs})
-
-    def test_remove(self):
-        key = self.underlying.get("file_1.mp3").album_key
-        songs = self.underlying._contents
-
-        # Remove all songs in Album 1
-        for i in range(1, 12, 3):
-            song = songs["file_%d.mp3" % i]
-            self.underlying.remove([song])
-
-        # Album 1 is all gone...
-        self.failUnlessEqual(self.library.get(key), None)
-
-        # ...but Album 2 is fine
-        key = self.underlying.get("file_2.mp3").album_key
-        album2 = self.library[key]
-        self.failUnlessEqual(album2.key, key)
-        self.failUnlessEqual(len(album2.songs), 4)
-
-    def test_misc(self):
-        # It shouldn't implement FileLibrary etc
-        self.failIf(getattr(self.library, "filename", None))
-
-
-class TAlbumLibrarySignals(TestCase):
-    def setUp(self):
-        lib = SongLibrary()
-        received = []
-
-        def listen(name, items):
-            received.append(name)
-
-        self._sigs = [
-            connect_obj(lib, 'added', listen, 'added'),
-            connect_obj(lib, 'changed', listen, 'changed'),
-            connect_obj(lib, 'removed', listen, 'removed'),
-        ]
-
-        albums = lib.albums
-        self._asigs = [
-            connect_obj(albums, 'added', listen, 'a_added'),
-            connect_obj(albums, 'changed', listen, 'a_changed'),
-            connect_obj(albums, 'removed', listen, 'a_removed'),
-        ]
-
-        self.lib = lib
-        self.albums = albums
-        self.received = received
-
-    def test_add_one(self):
-        self.lib.add([AlbumSong(1)])
-        self.failUnlessEqual(self.received, ["added", "a_added"])
-
-    def test_add_two_same(self):
-        self.lib.add([AlbumSong(1, "a1")])
-        self.lib.add([AlbumSong(5, "a1")])
-        self.failUnlessEqual(self.received,
-                             ["added", "a_added", "added", "a_changed"])
-
-    def test_remove(self):
-        songs = [AlbumSong(1, "a1"), AlbumSong(2, "a1"), AlbumSong(4, "a2")]
-        self.lib.add(songs)
-        self.lib.remove(songs[:2])
-        self.failUnlessEqual(self.received,
-                             ["added", "a_added", "removed", "a_removed"])
-
-    def test_change(self):
-        songs = [AlbumSong(1, "a1"), AlbumSong(2, "a1"), AlbumSong(4, "a2")]
-        self.lib.add(songs)
-        self.lib.changed(songs)
-        self.failUnlessEqual(self.received,
-                             ["added", "a_added", "changed", "a_changed"])
-
-    def tearDown(self):
-        for s in self._asigs:
-            self.albums.disconnect(s)
-        for s in self._sigs:
-            self.lib.disconnect(s)
-        self.lib.destroy()
 
 
 class Titer_paths(TestCase):
